pdf2word-function/requirements.txt
azure-functions
pdf2docx==0.5.8
PyMuPDF==1.23.26
python-docx==1.1.0